add-action-group (#6)
Add 'Add group' button to 'no action groups found' alert and to drop down selection when 1+ action groups already exist. When clicked, a dialogue prompts for the name of the new action group and, once confirmed by the user, the group is created and the selected inbox item is immediately moved to it.